+++ changelog.md
@@ -6,6 +6,7 @@
 
 * New: `Core` object has priority options for the loading actions
 * New: `Settings` object can specify some groups as network wide
+* New: settings now can render the `Clear` button for text-based field
 * New: added several more UI icons
 * Edit: improvements to the default styling for dashboard and tabs
 * Edit: improvements to base `Panel` object content file names